backend/requirements.txt
# Core LlamaIndex
llama-index-core==0.11.0
llama-index-embeddings-huggingface==0.3.0
llama-index-vector-stores-postgres==0.3.0

# Embeddings
sentence-transformers==2.7.0

# Web Framework
fastapi==0.115.0
uvicorn==0.30.0

# MCP
mcp==1.0.0

# Database
psycopg2-binary==2.9.9
pgvector==0.3.5
sqlalchemy==2.0.36

# Web & Document Processing
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.3.0
pypdf==4.3.1
python-docx==1.1.2
python-multipart==0.0.12

# SEO Tools
keybert==0.8.4
spacy==3.7.4
rake-nltk==1.0.6

# Utilities
pydantic==2.9.0
python-dotenv==1.0.1
